Generate sql with indirect queries.
