+++ pkg/sqlcache/sqltypes/types.go
@@ -25,11 +25,11 @@ const (
 
 // ListOptions represents the query parameters that may be included in a list request.
 type ListOptions struct {
-	ChunkSize  int
-	Resume     string
-	Filters    []OrFilter
-	SortList   SortList
-	Pagination Pagination
+	ChunkSize  int        `json:"chunkSize"`
+	Resume     string     `json:"resume"`
+	Filters    []OrFilter `json:"orFilters"`
+	SortList   SortList   `json:"sortList"`
+	Pagination Pagination `json:"pagination"`
 }
 
 // Filter represents a field to filter by.
@@ -40,12 +40,12 @@ type ListOptions struct {
 //
 // If more than one value is given for the `Match` field, we do an "IN (<values>)" test
 type Filter struct {
-	Field          []string
-	Matches        []string
-	Op             Op
-	Partial        bool
-	IsIndirect     bool
-	IndirectFields []string
+	Field          []string `json:"fields"`
+	Matches        []string `json:"matches"`
+	Op             Op       `json:"op"`
+	Partial        bool     `json:"partial"`
+	IsIndirect     bool     `json:"isIndirect"`
+	IndirectFields []string `json:"indirectFields"`
 }
 
 // OrFilter represents a set of possible fields to filter by, where an item may match any filter in the set to be included in the result.
@@ -59,20 +59,20 @@ type OrFilter struct {
 // The order is represented by prefixing the sort key by '-', e.g. sort=-metadata.name.
 // e.g. To sort internal clusters first followed by clusters in alpha order: sort=-spec.internal,spec.displayName
 type Sort struct {
-	Fields         []string
-	Order          SortOrder
-	IsIndirect     bool
-	IndirectFields []string
+	Fields         []string  `json:"fields"`
+	Order          SortOrder `json:"order"`
+	IsIndirect     bool      `json:"isIndirect"`
+	IndirectFields []string  `json:"indirectFields"`
 }
 
 type SortList struct {
-	SortDirectives []Sort
+	SortDirectives []Sort `json:"sortDirectives"`
 }
 
 // Pagination represents how to return paginated results.
 type Pagination struct {
-	PageSize int
-	Page     int
+	PageSize int `json:"pageSize"`
+	Page     int `json:"page"`
 }
 
 func NewSortList() *SortList {
